feat: detect mutation to values of reactive vars
A common source of error is mutation of values in reactive vars.
The reactive var cannot notice a change in its value (e.g. a list)
if the list is mutated in place.

A user can be mislead that it is working correctly, when another
reactive variable triggers a rerender, teaching bad habits.

Detecting mutations in Python without using proxies can be done by
making a deepcopy of the object, and comparing the reference to the
deepcopy.

This comes at the cost of performance and memory usage, therefore we
should enabled this by default in development mode, so app builders
are aware of mistakes while developing.
In production mode we can disable the mutation checks and behave as
before.